requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==6.7
Flask==1.1.1
idna==2.7
itsdangerous==0.24
Jinja2==2.10.3
MarkupSafe==0.23
plaid-python==3.4.0
python-http-client==3.2.1
requests==2.20.0
sendgrid==6.1.0
urllib3==1.24.3
Werkzeug==0.16.0
